requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
alembic==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrdict==2.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.10.0
binpickle==0.3.3
bleach==3.2.1
bokeh==2.4.2
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
certipy==0.1.3
chardet==3.0.4
click==7.1.2
cloudpickle==2.0.0
colorcet==3.0.0
conda==4.9.2
configparser==5.0.2
csr==0.2.0
cycler==0.10.0
dask==2022.3.0
dataclasses==0.6
datashader==0.13.0
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
distributed==2022.3.0
docker-pycreds==0.4.0
easy-dict==0.1.1
easydict==1.9
entrypoints==0.3
filelock==3.6.0
fsspec==2021.4.0
future==0.18.2
gdown==4.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
HeapDict==1.0.1
holoviews==1.14.8
huggingface-hub==0.4.0
imageio==2.9.0
ipdb==0.13.7
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-telemetry==0.1.0
jupyterhub==1.3.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lenskit==0.12.3
littleutils==0.2.2
llvmlite==0.35.0
lmdb==1.3.0
locket==0.2.1
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
multipledispatch==0.6.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
ogb==1.3.3
opencv-python==4.5.1.48
outdated==0.2.1
packaging==21.3
pamela==1.0.0
pandas==1.2.0
pandocfilters==1.4.3
panel==0.12.7
param==1.12.0
parso==0.8.1
partd==1.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.9
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyct==0.4.8
pyDeprecate==0.3.0
Pygments==2.7.3
pynndescent==0.5.6
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
pytorch-lightning==1.3.1
pytz==2020.5
pyviz-comms==2.1.0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2022.3.2
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.87
sacremoses==0.0.47
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
smmap==4.0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.3.22
subprocess32==3.5.4
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.8
tokenizers==0.11.6
toml==0.10.2
toolz==0.11.2
torch==1.10.2
torchaudio==0.7.0
torchmetrics==0.3.2
torchvision==0.11.3
tornado==6.1
traitlets==5.0.5
transformers==4.17.0
typing-extensions==4.1.1
umap-learn==0.5.2
wandb==0.10.30
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wilds==2.0.0
xarray==2022.3.0
yarl==1.6.3
zict==2.1.0
